+++ repos/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/pagination/KeyValuePaginationExtensions.kt
@@ -1,8 +1,10 @@
 package dev.inmo.micro_utils.repos.pagination
 
 import dev.inmo.micro_utils.pagination.*
+import dev.inmo.micro_utils.pagination.utils.getAllWithNextPaging
 import dev.inmo.micro_utils.repos.ReadStandardKeyValueRepo
 
+@Deprecated("Will be removed soon due to redundancy")
 suspend inline fun <Key, Value, REPO : ReadStandardKeyValueRepo<Key, Value>> REPO.doForAll(
     @Suppress("REDUNDANT_INLINE_SUSPEND_FUNCTION_TYPE")
     methodCaller: suspend REPO.(Pagination) -> PaginationResult<Key>,
@@ -15,17 +17,17 @@ suspend inline fun <Key, Value, REPO : ReadStandardKeyValueRepo<Key, Value>> REP
     }
 }
 
+@Deprecated("Will be removed soon due to redundancy")
 suspend inline fun <Key, Value, REPO : ReadStandardKeyValueRepo<Key, Value>> REPO.doForAll(
     block: (List<Pair<Key, Value>>) -> Unit
 ) = doForAll({ keys(it, false) }, block)
 
 suspend inline fun <Key, Value, REPO : ReadStandardKeyValueRepo<Key, Value>> REPO.getAll(
     @Suppress("REDUNDANT_INLINE_SUSPEND_FUNCTION_TYPE")
-    methodCaller: suspend REPO.(Pagination) -> PaginationResult<Key>
-): List<Pair<Key, Value>> {
-    val resultList = mutableListOf<Pair<Key, Value>>()
-    doForAll(methodCaller) {
-        resultList.addAll(it)
-    }
-    return resultList
+    crossinline methodCaller: suspend REPO.(Pagination) -> PaginationResult<Key>
+): List<Pair<Key, Value>> = getAllWithNextPaging {
+    val result = methodCaller(it)
+    result.changeResultsUnchecked(
+        result.results.mapNotNull { it to (get(it) ?: return@mapNotNull null) }
+    )
 }